refactor(journal): hoist shadow admission to the journal parent
The delivery-journal shadow admission gate (mode + pool + cohort) lived in
`journal::watcher` while the watcher was the only family that could not reach
the sink's `begin_fresh`. T1 S4 of issue 5071 adds the turn_bridge/controller
family, which needs the identical gate, so it moves to the shared parent for
the same reason `PROCESS_OBSERVER` did in S3a.

Move-only: the two watcher call sites now spell it `super::admit`, the
`DeliveryJournalMode` import follows the function, and no predicate changes.

The delivery journal is shadow-only and has no reader: a row exists only when
a PG pool, `DeliveryJournalMode::Shadow` and the cohort selection all agree, so
nothing here can influence what production delivers.
